% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scBio.R
\docType{data}
\name{BulkFlu}
\alias{BulkFlu}
\title{Gene expression profiles of flu and pbs sample.}
\format{A matrix with 1858 rows (genes) and 74 columns (samples).}
\usage{
BulkFlu
}
\description{
A dataset containing the RNA-seq profiles of colaborative-cross (CC) mice 2 days after the infection with either the flu virus or PBS.
}
\keyword{datasets}
