% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scBio.R
\docType{data}
\name{SCCellSpace}
\alias{SCCellSpace}
\title{Single-cell cell space.}
\format{A matrix with 349 rows (cells) and 2 columns (dimensions).}
\usage{
SCCellSpace
}
\description{
A dataset containing a 2-dim cell space of all single-cells in the SCFlu dataset.
}
\keyword{datasets}
