% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scBio.R
\docType{data}
\name{SCFlu}
\alias{SCFlu}
\title{Gene expression profiles of lung cells after influenza infection.}
\format{A matrix with 1858 rows (genes) and 349 columns (cells).}
\usage{
SCFlu
}
\description{
A dataset containing the RNA-seq profiles of lung cells from multiple cell types, taken from two mice 2 days after the infection with either the flu virus or PBS.
}
\keyword{datasets}
