% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scBio.R
\docType{data}
\name{SCLabels}
\alias{SCLabels}
\title{Single-cell classification into cell types.}
\format{A vector with 349 values.}
\usage{
SCLabels
}
\description{
A dataset containing the classification of each cell in the SCFlu dataset to a specific cell type.
}
\keyword{datasets}
